+++ fixtures/drift_lateral/fail_lateral_new/a/a.h
@@ -1,0 +1,3 @@
+#pragma once
+#include "b/b.h"
+#include "utils/u.h"