+++ fixtures/drift_lateral/fail_lateral_sdp/b/b.h
@@ -1,0 +1,3 @@
+#pragma once
+#include "c/c.h"
+#include "d/d.h"